Restarting once more, this time with making issues on the repo project board
