--- assignment02/weatherapp/src/main/resources/weather_cache.json
@@ -1,58 +1,0 @@
-[{
-        "id": "IDS60901",
-        "name": "Adelaide (West Terrace / ngayirdapira)",
-        "state": "SA",
-        "time_zone": "CST",
-        "lat": -34.9,
-        "lon": 138.6,
-        "local_date_time": "21/05:00pm",
-        "local_date_time_full": "20230921170000",
-        "air_temp": 13.3,
-        "apparent_t": 9.5,
-        "cloud": "Partly cloudy",
-        "dewpt": 5.7,
-        "press": 1023.9,
-        "rel_hum": 60,
-        "wind_dir": "S",
-        "wind_spd_kmh": 15,
-        "wind_spd_kt": 8
-    },
-    {
-        "id": "IDS60902",
-        "name": "Melbourne",
-        "state": "VIC",
-        "time_zone": "AEDT",
-        "lat": -37.8,
-        "lon": 144.9,
-        "local_date_time": "21/05:30pm",
-        "local_date_time_full": "20240921173000",
-        "air_temp": 14.0,
-        "apparent_t": 12.0,
-        "cloud": "Clear",
-        "dewpt": 6.0,
-        "press": 1024.0,
-        "rel_hum": 55,
-        "wind_dir": "NW",
-        "wind_spd_kmh": 10,
-        "wind_spd_kt": 5
-    },
-    {
-        "id": "IDS60903",
-        "name": "Sydney",
-        "state": "NSW",
-        "time_zone": "AEDT",
-        "lat": -33.9,
-        "lon": 151.2,
-        "local_date_time": "21/05:30pm",
-        "local_date_time_full": "20240921173000",
-        "air_temp": 18.0,
-        "apparent_t": 17.0,
-        "cloud": "Sunny",
-        "dewpt": 8.0,
-        "press": 1015.0,
-        "rel_hum": 50,
-        "wind_dir": "E",
-        "wind_spd_kmh": 20,
-        "wind_spd_kt": 10
-    }
-]
